requirements.txt
azure-cognitiveservices-vision-customvision>=3.1.0
gradio>=3.50.2
numpy>=1.24.3
opencv-python>=4.8.0
Pillow>=10.0.0
pandas>=2.0.3
jupyter>=1.0.0
matplotlib>=3.7.2 